+++ types/wlr_xcursor_manager.c
@@ -46,7 +46,8 @@ bool wlr_xcursor_manager_load(struct wlr_xcursor_manager *manager,
 		return false;
 	}
 	theme->scale = scale;
-	theme->theme = wlr_xcursor_theme_load(manager->name, manager->size * scale);
+	theme->theme = wlr_xcursor_theme_load(manager->name,
+		(int)(manager->size * scale));
 	if (theme->theme == NULL) {
 		free(theme);
 		return false;
